backend/requirements.txt
# Updated for Railway cache bust - March 2026
annotated-doc==0.0.4
annotated-types==0.7.0
openai==2.32.0
anyio==4.12.1
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
deprecation==2.1.0
distro==1.9.0
fastapi==0.115.0
filelock==3.25.0
fsspec==2026.2.0
gotrue==2.12.4
h11==0.16.0
h2==4.3.0
hf-xet==1.3.2
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.2
huggingface_hub==1.6.0
hyperframe==6.1.0
idna==3.11
jiter==0.13.0
markdown-it-py==4.0.0
mdurl==0.1.2
packaging==26.0
postgrest==0.16.11
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.3
realtime==2.28.0
rich==14.3.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.38.6
storage3==0.7.7
StrEnum==0.4.15
supabase==2.7.4
supafunc==0.5.1
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.30.6
watchfiles==1.1.1
websockets==15.0.1
xmltodict==0.13.0
